Add more OOP features to source code
All of the code is currently stored in a single class 'Blob' with the use of nested classes (e.g. Task, Deadline, Todo...).

Separation of Nested Classes into separate files. Creation of more classes such as Ui, Storage, Parser and TaskList.

Use of separate files for the different varying classes makes it much easier for the developer to isolate each portion of the code. Addition of more classes (such as Ui, Storage, Parser, and TaskList) help to abstract away more code for easier understanding and debugging.
