+++ src/main/java/Todo.java
@@ -1,0 +1,12 @@
+public class Todo extends Task{
+    public Todo(String name, boolean isDone) {
+        super(name,isDone);
+        super.type = "T";
+    }
+
+    @Override
+    public String toString() {
+        return "[T]" + "[" + this.check() + "] " + this.name;
+    }
+}
+
