+++ src/main/java/database.csv
@@ -1,2 +1,2 @@
 type,is_checked,task_name,time1,time2
-D,0,eat,2024-08-26T14:30
+E,0,CS2030S Lab,2024-08-26T14:00,2024-08-26T16:00
